// migrations/snapshots/cca0ba04d7fca33f67d5f0443a2fe0cb49a3f02e5da2582f7d000221caac53ae/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Acronym: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Meaning: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly acronymId: CodecTypes['pg/int4@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly normalizedText: CodecTypes['pg/text@1']['output'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly ScoreTransaction: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly acronymId: CodecTypes['pg/int4@1']['output'];
      readonly meaningId: CodecTypes['pg/int4@1']['output'];
      readonly amount: CodecTypes['pg/int4@1']['output'];
      readonly type: 'NEW_ACRONYM' | 'EXISTING_ENTRY' | 'DUPLICATE_FOUND';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly externalId: CodecTypes['pg/text@1']['output'];
      readonly displayName: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly score: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Acronym: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Meaning: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly acronymId: CodecTypes['pg/int4@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly normalizedText: CodecTypes['pg/text@1']['input'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly ScoreTransaction: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly acronymId: CodecTypes['pg/int4@1']['input'];
      readonly meaningId: CodecTypes['pg/int4@1']['input'];
      readonly amount: CodecTypes['pg/int4@1']['input'];
      readonly type: 'NEW_ACRONYM' | 'EXISTING_ENTRY' | 'DUPLICATE_FOUND';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly externalId: CodecTypes['pg/text@1']['input'];
      readonly displayName: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly score: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly acronym: {
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly meaning: {
      readonly acronymId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly normalizedText: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly scoreTransaction: {
      readonly acronymId: CodecTypes['pg/int4@1']['output'];
      readonly amount: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly meaningId: CodecTypes['pg/int4@1']['output'];
      readonly type: 'NEW_ACRONYM' | 'EXISTING_ENTRY' | 'DUPLICATE_FOUND';
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly displayName: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly externalId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly score: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly acronym: {
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly meaning: {
      readonly acronymId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdByUserId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly normalizedText: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly scoreTransaction: {
      readonly acronymId: CodecTypes['pg/int4@1']['input'];
      readonly amount: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly meaningId: CodecTypes['pg/int4@1']['input'];
      readonly type: 'NEW_ACRONYM' | 'EXISTING_ENTRY' | 'DUPLICATE_FOUND';
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly displayName: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly externalId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly score: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;